refactor(dialog): DialogManager with XamlRoot, split into clean files
Dialog/DialogManager.cs — orchestrator init with FormMain, Show/ShowAsync,
  ShowText/ShowMarkdown/ShowInput/ShowSelect conveniences, converter logic.
Dialog/DialogContext.cs, DialogButton.cs, DialogResult.cs, DialogTheme.cs
  — type files (moved from PCL.Core.UI).
Dialog/LegacyDialog.cs — partial ModMain: MyMsgBoxConverter, WaitingMyMsgBox,
  thin wrappers delegating to DialogManager.Instance.
PCL.Core/UI/MsgBoxWrapper.cs — reverted to event pattern, no Dialog dependency.
Removed: PCL.Core/UI/Dialog.cs, Modules/ModMain.Dialog.cs (merged above).
ModMain.cs timer calls DialogManager.Instance.Tick().
