+++ Plain Craft Launcher 2/Dialog/DialogResult.cs
@@ -1,0 +1,9 @@
+namespace PCL.Core.UI;
+
+public static class DialogResult
+{
+    public const int Ok = 1;
+    public const int Cancel = 2;
+    public const int Yes = 6;
+    public const int No = 7;
+}
