+++ Plain Craft Launcher 2/Dialog/DialogTheme.cs
@@ -1,0 +1,8 @@
+namespace PCL.Core.UI;
+
+public enum DialogTheme
+{
+    Info,
+    Warning,
+    Error
+}
